Convert experimental completion into options
`globus --experimental-completer`,
`globus --experimental-bash-completer`,
and
`globus --experimental-zsh-completer`

do the same work as before. Minor refactoring for layout.
